Support for SAMD21 family: Add sam/d21 family to Makefile
Because SAMD10 and SAMD21 are families like STM32F0 and STM32F1,
I decided that a family 'sam/d' is not sufficient.

The GPIO module and interrupt vectors of SAMD09, 10, and 11 are compatible,
but SAMD20 and 21 are incompatible.
So it is necessary to write independent irq.json files,
which is easiest by making a new family.
It also does not make sense to have a doxygen module for GPIO for all SAMD,
because SAMD21 uses PORT instead.